requirements.txt
streamlit==1.32.0
openai==1.12.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
supabase-py==2.3.0 